LayoutTests/http/tests/websocket/tests/hybi/reserved-bits.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
description("Test whether WebSocket rejects frames whose reserved bit is on.");

window.jsTestIsAsync = true;
if (window.layoutTestController)
    layoutTestController.overridePreference("WebKitHixie76WebSocketProtocolEnabled", 0);

var closeEvent;

// bitNumber must be 1, 2, or 3.
function doTest(bitNumber)
{
    var url = "ws://localhost:8880/websocket/tests/hybi/reserved-bits?bit=" + bitNumber;
    var ws = new WebSocket(url);

    debug("Bit " + bitNumber + ": Test started.");

    ws.onopen = function()
    {
        debug("onopen() was called.");
    };

    ws.onmessage = function(event)
    {
        var message = event.data;
        testFailed("onmessage() was called. (message = \"" + message + "\")");
    };

    ws.onclose = function(event)
    {
        debug("onclose() was called.");
        closeEvent = event;
        shouldBeFalse("closeEvent.wasClean");
        if (bitNumber === 3)
            finishJSTest();
        else
            doTest(bitNumber + 1);
    };
}

doTest(1);

</script>
<script src="../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
